requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
anthropic==0.26.0
pdfplumber==0.11.0
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.9
gtts==2.5.1
httpx==0.27.0
aiofiles==23.2.1
